Reorganizing plugin structure.
Moving stuff around to more closely match Zeek's standard layout that
init-plugin creates.
