+++ include/file-analyzer.h
