+++ include/zeek-compat.h
