--- plugin/debug.h
@@ -1,6 +1,0 @@
-// Copyright (c) 2020-2021 by the Zeek Project. See LICENSE for details.
-
-#pragma once
-
-// TODO: We should separate out the debug support here vs from compiler/.
-#include "../compiler/debug.h"
