--- plugin/include/debug.h
@@ -1,1 +1,0 @@
-../../compiler/debug.h
