--- plugin/scripts/__preload__.zeek
@@ -1,2 +1,0 @@
-@load ../lib/bif
-@load ./base/spicy
