--- plugin/scripts/base/spicy/__load__.zeek
@@ -1,1 +1,0 @@
-@load ./main
