+++ scripts/Zeek/Spicy/misc/record-spicy-batch.zeek
