+++ scripts/__load__.zeek
@@ -1,0 +1,1 @@
+@load Zeek/Spicy/bare.zeek
