+++ spicy/zeek.spicy
