+++ spicy/zeek_file.spicy
