+++ src/compiler/README
@@ -1,6 +1,5 @@
 This library implements all code generation & compilation for
 interfacing Spicy with Zeek. It's independent of having a Zeek process
 itself available, everything that requires having access to Zeek
-itself is out of scope. This way, this library will later be able to
-support a standalone compiler tool that outputs C++ code for static
-compilation.
+itself is out of scope. This way, this library is able to
+support a standalone compiler tool as well; that's `spicyz`.
