+++ src/consts.bif
