+++ src/file-analyzer.cc
